backend_flask/requirements.txt
Flask
python-dotenv
facenet-pytorch
torch
torchvision
Pillow